requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.6
django-celery-results==2.5.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
h11==0.14.0
idna==3.7
kombu==5.3.7
outcome==1.3.0.post0
prompt_toolkit==3.0.46
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
redis==5.0.5
requests==2.32.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
